Revert "Upgrade to medium"
