+++ csgame/nlp_loader.py
@@ -1,7 +1,7 @@
-from en_core_web_md import load
+import en_core_web_sm
 import spacy
 
-nlp = load()  # Single load in memory till killed
+nlp = en_core_web_sm.load()  # Single load in memory till killed
 
 if not isinstance(nlp, spacy.lang.en.English):
     raise TypeError("Model given is not of type {}.".format("spacy.lang.en.English"))
